add db layer

+++ models/ColumnModel.go
@@ -1,8 +1,8 @@
 package models
 
 type ColumnModel struct {
-	Id        uint   `json:"id"`
+	Id        int64  `json:"id"`
 	Name      string `json:"name" validate:"required,min=1,max=255"`
 	Position  uint   `json:"position"`
-	ProjectId uint   `json:"project_id"`
+	ProjectId int64  `json:"project_id"`
 }
