+++ models/CommentModel.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type CommentModel struct {
-	Id        uint      `json:"id"`
+	Id        int64     `json:"id"`
 	Text      string    `json:"text" validate:"required,min=1,max=5000"`
 	CreatedAt time.Time `json:"created_at"`
-	TaskId    uint      `json:"task_id"`
+	TaskId    int64     `json:"task_id"`
 }
